Changes: + Frame Buffer Object class, use Begin and End to render to a texture buffer. + Vertex Buffer Object class, this is used by the Mesh loader and will be used for creating terrain. + Mesh class, loads a .obj from wavefront. + Shader class, loads a shader program. Call Activate, render what you want affected by the shader and then Deactivate. + glInfo, this is just so that we can check what hardware functions we have available to us. Can be ignored. > Modified ResourceManager so that you can load shaders and mesh using it. + Windows RC file, it simply gives the exe a .ico > Changed the SceneTask so it does not start each scene with a DRIVEN camera. + Added key 'p' to Keyboard in SceneTask. This allows you to view the camera trajectory splines.
Bugs:

None that I am aware of.

ToDo:

Add the terrain class. Load a texture file as a heightmap to a terrain class and render it using a VBO. Possibly stick it into a quad tree for efficiency.

Notes:

Incase I never mentioned before. The keys are as follows:

W,A,S,D,Q,E - Move the camera
Space - Puts camera into DRIVEN, where it follows a trajectory which is hardcoded in the SimpleScene but can be loaded or saved to a text file.
M - Show/Hide mouse cursor
P - Show/Hide splines
Esc - Quit
+ or - on Number Pad - Increase / Decrease speed for inputs or global timer updates which happen in SceneTask before SceneManager.Idle() is called.