NOTE:
Uploading for you guys to get a start on implementing networking, Kinect and GUI? Spent way too much time on it this week!

CHANGES:

CPointer - Pointer class used to select a point on the terrain. Render a little thimble object which changes size depending on the Area of Influence (Area which it is affecting)

Player - A temporary player class. Used to pass the pointer logic to the pointer and a rough idea for a networking client player.

terrain editing logic.png - This is an image I created to work out how to edit the terrain.

Terrain.Edit(TYPE, vec2 Position, int AoISize) - Created this function to edit the terrain, TYPE is what you are editing i.e. HEIGHTMAP, GRASS, ROCK etc. Position is the center point of the AoI, this is the currently selected vertice(or pixel, depends how you look at it). AoISize, is the radius of the area of influence. This should be passed in from the players pointer m_AoISize. 
Note: Only HeightMap editing currently implemented.

This is just a rough overview, please have a look at the code to understand what is actually happening.

BUGS:

- Terrain editing logic is a little buggy, esoecially when you reach the bounding box height limit. It forces all vertices under the AoI to BBox.max.y.
- Terrain shrinking currently ignores certain pixels because their is no buffer for what vertices are currently selected.
- If you try to edit a vertice which doesnt exist, i.e. outside the BBox, then it will give an access violation.
- If you the terrain chunk isn't in the view frustum it isnt getting rendered after editing.
- Terrain edits are only affecting the VBO vertice positions and not the tangents, normals and texture coords.

TODO:
- Use an interopalation function(cubic) for height editing to make it smoother, atm it is linearly interpolated between center point and radius.
- Stop it from trying to edit vertices that dont exist, i.e. check if it is in the BBox.
- Save all editing to a texture file, i.e. heightmap and diffuse map.
- Fix the terrain shrinking by adding padding to the vertice selection.
- Change the way the player moves the pointer so that if the pointer isn't in the view frustum then we move it into the view frustum. Also needs to be change the movement so that it is parallel with the camera and not axis aligned.
- Fix terrain not rendering if it has been moved and it isn't in the view frustum.
- Editing terrain also edits the VBO, tangent, normal and texture coords.
- Add some pretty shaders.
- Add Kinect.
- Add networking.